// src/generated/plugins/authcrunch-ids.ts
// Code generated by tygo. DO NOT EDIT.

//////////
// source: config.go

/**
 * IdentityStoreConfig represents an identity store configuration.
 */
export interface IdentityStoreConfig {
  name?: string;
  kind?: string;
  params?: { [key: string]: any };
}

//////////
// source: method.go

/**
 * IdentityStoreType identifies identity store type.
 */
export type IdentityStoreType = number /* int */;
/**
 * UNKNOWN is unknown identity store type.
 */
export const UNKNOWN: IdentityStoreType = 0;
/**
 * LOCAL identifies local identity store.
 */
export const LOCAL: IdentityStoreType = 1;
/**
 * LDAP identifies local LDAP identity store.
 */
export const LDAP: IdentityStoreType = 2;

//////////
// source: store.go

/**
 * IdentityStore represents identity store.
 */
export type IdentityStore = any;