// src/generated/plugins/authcrunch-saml.ts
// Code generated by tygo. DO NOT EDIT.

import type { LoginIcon } from "./authcrunch-icons";

//////////
// source: config.go

/**
 * Config holds the configuration for the IdentityProvider.
 */
export interface Config {
  /**
   * Name is the unique name associated with the IdentityProvider.
   */
  name?: string;
  /**
   * Realm the authentication realm associated with the IdentityProvider.
   */
  realm?: string;
  /**
   * Driver is the name of the driver associated with the IdentityProvider, e.g. azure.
   */
  driver?: string;
  /**
   * IdpMetadataLocation is the path to the Identity Provider metadata.
   */
  idp_metadata_location?: string;
  /**
   * IdpSignCertLocation is the path to the Identity Provider signing certificate.
   */
  idp_sign_cert_location?: string;
  /**
   * IdpLoginURL is the SAML authentication endpoint with the Identity Provider.
   */
  idp_login_url?: string;
  /**
   * TenantID is the tenant ID associated with the IdentityProvider.
   */
  tenant_id?: string;
  /**
   * ApplicationID is the application ID associated with the IdentityProvider.
   */
  application_id?: string;
  /**
   * ApplicationName  is the application name associated with the IdentityProvider.
   */
  application_name?: string;
  /**
   * EntityID is the "Identifier (Entity ID)" an administrator
   * specifies in "Set up Single Sign-On with SAML" in Azure AD
   * Enterprise Applications.
   */
  entity_id?: string;
  /**
   * AssertionConsumerServiceURLs is the list of URLs server instance is listening
   * on. These URLs are known as SP Assertion Consumer Service endpoints. For
   * example, users may access a website via http://app.domain.local. At the
   * same time the users may access it by IP, e.g. http://10.10.10.10. or
   * by name, i.e. app. Each of the URLs is a separate endpoint.
   */
  acs_urls?: string[];
  tls_insecure_skip_verify?: boolean;
  /**
   * LoginIcon is the UI login icon attributes.
   */
  login_icon?: LoginIcon;
}

//////////
// source: provider.go

/**
 * IdentityProvider represents SAML-based identity provider.
 */
export interface IdentityProvider {}