// src/generated/plugins/authcrunch-kms.ts
// Code generated by tygo. DO NOT EDIT.

//////////
// source: config.go

/**
 * CryptoKeyConfig is common token-related configuration settings.
 */
export interface CryptoKeyConfig {
  /**
   * Seq is the order in which a key would be processed.
   */
  seq?: number /* int */;
  /**
   * ID is the key ID, aka kid.
   */
  id?: string;
  /**
   * Usage is the intended key usage. The values are: sign, verify, both,
   * or auto.
   */
  usage?: string;
  /**
   * TokenName is the token name associated with the key.
   */
  token_name?: string;
  /**
   * Source is either config or env.
   */
  source?: string;
  /**
   * Algorithm is either hmac, rsa, or ecdsa.
   */
  algorithm?: string;
  /**
   * EnvVarName is the name of environment variables holding either the value of
   * a key or the path a directory or file containing a key.
   */
  env_var_name?: string;
  /**
   * EnvVarType indicates how to interpret the value found in the EnvVarName. If
   * it is blank, then the assumption is the environment variable value
   * contains either public or private key.
   */
  env_var_type?: string;
  /**
   * EnvVarValue is the value associated with the environment variable set by EnvVarName.
   */
  env_var_value?: string;
  /**
   * FilePath is the path of a file containing either private or public key.
   */
  file_path?: string;
  /**
   * DirPath is the path to a directory containing crypto keys.
   */
  dir_path?: string;
  /**
   * TokenLifetime is the expected token grant lifetime in seconds.
   */
  token_lifetime?: number /* int */;
  /**
   * Secret is the shared key used with HMAC algorithm.
   */
  token_secret: string;
  /**
   * PreferredSignMethod is the preferred method to sign tokens, e.g.
   * all HMAC keys could use HS256, HS384, and HS512 methods. By default,
   * the preferred method is HS512. However, one may prefer using HS256.
   */
  token_sign_method?: string;
  /**
   * EvalExpr is a list of expressions evaluated whether a specific key
   * should be used for signing and verification.
   */
  token_eval_expr: string[];
}

//////////
// source: key.go

/**
 * CryptoKey contains a crypto graphic key and associated metadata.
 */
export interface CryptoKey {
  config?: CryptoKeyConfig;
  sign?: CryptoKeyOperator;
  verify?: CryptoKeyOperator;
}
/**
 * CryptoKeyTokenOperator represents CryptoKeyOperator token operator.
 */
export interface CryptoKeyTokenOperator {
  id?: string;
  name?: string;
  max_lifetime?: number /* int */;
  methods?: { [key: string]: any };
  preferred_methods?: string[];
  default_method?: string;
  capable?: boolean;
}
/**
 * CryptoKeyOperator represents CryptoKey operator.
 */
export interface CryptoKeyOperator {
  token?: CryptoKeyTokenOperator;
  secret?: any;
  capable?: boolean;
}

//////////
// source: keystore.go

/**
 * CryptoKeyStore constains keys assembled for a specific purpose, i.e. signing or
 * validation.
 */
export interface CryptoKeyStore {}