// src/generated/plugins/authcrunch-acl.ts
// Code generated by tygo. DO NOT EDIT.

//////////
// source: acl.go

/**
 * AccessList is a collection of access list rules.
 */
export interface AccessList {}

//////////
// source: condition.go

//////////
// source: rule.go

/**
 * RuleConfiguration consists of a list of conditions and and actions
 */
export interface RuleConfiguration {
  comment?: string;
  conditions?: string[];
  action?: string;
}